fix healthMultiplier can only be a full number

+++ src/main/java/de/melanx/boohoo/ModConfig.java
@@ -14,7 +14,7 @@ public class ModConfig {
     public static double spawnChance = 0.001;
 
     @Config("The multiplier for calculating the additional health for a ghost")
-    public static int healthMultiplier = 1;
+    public static double healthMultiplier = 1;
 
     @Config("Duration in ticks the ghost will stay in the world after the player died.")
     @IntRange(max = Integer.MAX_VALUE - 1)
